[generator] New structure python projects. Refactored booking.
