+++ tools/python/booking/api/exceptions.py
@@ -1,0 +1,14 @@
+class BookingError(Exception):
+    pass
+
+
+class HTTPError(BookingError):
+    pass
+
+
+class AttemptsSpentError(BookingError):
+    pass
+
+
+class GettingMinPriceError(BookingError):
+    pass
