+++ tools/python/booking/requirements.txt
@@ -1,0 +1,4 @@
+eviltransform
+ratelimit
+requests
+tqdm
